backend/requirements.txt
# Django core
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.0

# OAuth & Social Authentication
django-allauth==0.57.0
requests==2.32.4
PyJWT==2.10.1
cryptography==45.0.5

# Database
psycopg2-binary==2.9.10

# WebSocket support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
websockets==15.0.1

# CORS for frontend
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Dev dependencies
django-debug-toolbar==4.2.0

# Testing
pytest-django==4.5.2
pytest==7.4.3

# Production
gunicorn==21.2.0
